WIP: animated gif support
I'm sure this code patch is correct way to handle the gif file in csharp, but, Don't know why, It does not work now. BTW, after uploading a gif image directly through the streamdeck menu, the h/w device shows the animation, but not shown in the SD software.
Hence, I suspect something is wrong with the SD software, not this code.